Added auth0 support, removed authService project

+++ src/backend/Playprism/Services/TournamentService/Playprism.Services.TournamentService.API/Controllers/TournamentController.cs
@@ -1,27 +1,32 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using Playprism.Services.TournamentService.BLL.Interfaces;
-using Playprism.Services.TournamentService.DAL.Entities;
-using System.Threading.Tasks;
+﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using Playprism.Services.TournamentService.BLL.Dtos;
 using Playprism.Services.TournamentService.BLL.Exceptions;
+using Playprism.Services.TournamentService.BLL.Interfaces;
 using Playprism.Services.TournamentService.BLL.Interfaces.CompetitionOrganizer;
+using Playprism.Services.TournamentService.DAL.Entities;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using Playprism.Services.TournamentService.BLL.Dtos;
+using System.Security.Claims;
+using System.Threading.Tasks;
 
 namespace Playprism.Services.TournamentService.API.Controllers
 {
+    [Authorize]
     [ApiController]
     [Route("api/v1/[controller]")]
     public class TournamentController : ControllerBase
     {
         private readonly ITournamentService _tournamentService;
         private readonly ICompetitionProcess _competitionProcess;
+        private readonly IAuthorizationService _authorizationService;
 
-        public TournamentController(ITournamentService tournamentService, ICompetitionProcess competitionProcess)
+        public TournamentController(ITournamentService tournamentService, ICompetitionProcess competitionProcess, IAuthorizationService authorizationService)
         {
             _tournamentService = tournamentService;
             _competitionProcess = competitionProcess;
+            _authorizationService = authorizationService;
         }
 
         [HttpGet]
@@ -54,6 +59,7 @@ namespace Playprism.Services.TournamentService.API.Controllers
                 return BadRequest();
             }
 
+            entity.OwnerId = User.FindFirst(ClaimTypes.NameIdentifier).Value;
             var result = await _tournamentService.AddTournamentAsync(entity);
             return Accepted(result);
         }
@@ -73,6 +79,13 @@ namespace Playprism.Services.TournamentService.API.Controllers
                 return BadRequest();
             }
 
+            var tournament = _tournamentService.GetTournamentAsync(id);
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, tournament, "TournamentOwnerPolicy");
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
+
             try
             {
                 await _tournamentService.UpdateTournamentAsync(entity);
@@ -89,6 +102,12 @@ namespace Playprism.Services.TournamentService.API.Controllers
         [ProducesResponseType(StatusCodes.Status202Accepted)]
         public async Task<ActionResult> Delete(int id)
         {
+            var tournament = _tournamentService.GetTournamentAsync(id);
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, tournament, "TournamentOwnerPolicy");
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
             await _tournamentService.DeleteTournamentAsync(id);
             return Accepted();
         }
@@ -106,6 +125,7 @@ namespace Playprism.Services.TournamentService.API.Controllers
             // todo: playerId/participantId has to be get from JWT
             try
             {
+                participantEntity.UserId = User.FindFirst(ClaimTypes.NameIdentifier).Value;
                 await _tournamentService.AddNewParticipantAsync(participantEntity);
             }
             catch (EntityNotFoundException)
@@ -120,6 +140,11 @@ namespace Playprism.Services.TournamentService.API.Controllers
         public async Task<ActionResult> StartTournament(int id)
         {
             var tournament = await _tournamentService.GetTournamentAsync(id);
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, tournament, "TournamentOwnerPolicy");
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
             await _competitionProcess.GenerateBracketAsync(tournament);
 
             return Ok();
@@ -128,8 +153,14 @@ namespace Playprism.Services.TournamentService.API.Controllers
         [HttpPost("{id}/rounds/{roundId}/close")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        public async Task<ActionResult> CloseRound(int roundId)
+        public async Task<ActionResult> CloseRound(int id, int roundId)
         {
+            var tournament = await _tournamentService.GetTournamentAsync(id);
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, tournament, "TournamentOwnerPolicy");
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
             try
             {
                 await _competitionProcess.FinishRoundAsync(roundId);
@@ -145,6 +176,12 @@ namespace Playprism.Services.TournamentService.API.Controllers
         [ProducesResponseType(typeof(BracketResponse), StatusCodes.Status200OK)]
         public async Task<ActionResult<BracketResponse>> GetBracket(int id)
         {
+            var tournament = await _tournamentService.GetTournamentAsync(id);
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, tournament, "TournamentOwnerPolicy");
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
             var bracket = await _competitionProcess.GenerateResponseBracketAsync(id);
             return Ok(bracket);
         }
